I have no idea what am i writing

+++ src/structures/Node.ts
@@ -302,13 +302,14 @@ export class Node {
 
 				if (!this.manager.options.autoResume) return;
 
-				this.rest.patch(`/v4/sessions/${this.sessionId}`, {
+				await this.rest.patch(`/v4/sessions/${this.sessionId}`, {
 					resuming: true,
 					timeout: 360,
 				});
 
 				const resumedPlayers = <any[]>await this.rest.getAllPlayers();
 				for (const resumedPlayer of resumedPlayers) {
+					if (this.manager.players["get"](resumedPlayer.guildId)) return;
 					const previousInfosPlayer: any = this.manager.db.get(`players.${resumedPlayer.guildId}`) || {};
 					if (!previousInfosPlayer.guild || !previousInfosPlayer.voiceChannel || !previousInfosPlayer.textChannel) {
 						this.manager.db.delete(`players.${resumedPlayer.guildId}`);
@@ -325,28 +326,31 @@ export class Node {
 						data: previousInfosPlayer.data ?? {},
 					});
 
-					if (player.state !== "CONNECTED") player.connect();
+					player.connect();
 					if (!previousInfosPlayer.current) return;
 					player.state = "RESUMING";
 
 					let decoded = await this.manager.decodeTrack(previousInfosPlayer.current);
 					player.queue.current = TrackUtils.build(decoded);
 
-					if (resumedPlayer.filters.distortion) player.filters.distortion = resumedPlayer.filters.distortion;
-					if (resumedPlayer.filters.equalizer) player.filters.equalizer = resumedPlayer.filters.equalizer;
-					if (resumedPlayer.filters.karaoke) player.filters.karaoke = resumedPlayer.filters.karaoke;
-					if (resumedPlayer.filters.rotation) player.filters.rotation = resumedPlayer.filters.rotation;
-					if (resumedPlayer.filters.timescale) player.filters.timescale = resumedPlayer.filters.timescale;
-					if (resumedPlayer.filters.vibrato) player.filters.vibrato = resumedPlayer.filters.vibrato;
-					if (resumedPlayer.filters.volume) player.filters.volume = resumedPlayer.filters.volume;
-					player.position = resumedPlayer.state.position;
-					player.playing = true;
-					player.paused = false;
-
 					previousInfosPlayer.queue.map(async (encoded: string) => {
 						decoded = await this.manager.decodeTrack(encoded);
 						player.queue.add(TrackUtils.build(decoded));
 					});
+
+					await player.node.rest.updatePlayer({
+						guildId: resumedPlayer.guildId,
+						data: {
+							encodedTrack: decoded.encoded,
+							volume: resumedPlayer.volume,
+							position: resumedPlayer.state.position,
+							filters: resumedPlayer.filters,
+						},
+					});
+
+					player.playing = true;
+					player.paused = false;
+					player.position = resumedPlayer.state.position;
 
 					setTimeout(() => (player.state = "CONNECTED"), 5000);
 				}
